pyproject.toml
[project]
name = "bds-agent"
version = "2.1.0"
description = "AI-powered Real Estate Search Agent for Vietnamese Market"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = {text = "MIT"}

authors = [
    {name = "BDS Agent Team"}
]

keywords = ["real-estate", "crawler", "vietnam", "ai", "fastapi"]

dependencies = [
    # Core Web Scraping
    "crawl4ai[all]>=0.3.74",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "httpx>=0.26.0",

    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=2.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Vector Database
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",

    # LLM
    "langchain-groq>=0.1.9",
    "langchain-core>=0.1.52",

    # Scheduler
    "apscheduler>=3.10.4",

    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.3",
    "cachetools>=5.3.0",

    # Data Processing
    "pandas>=2.1.0",
    "phonenumbers>=8.13.0",
    "python-dateutil>=2.8.2",

    # Logging & Observability
    "structlog>=24.1.0",
    "python-json-logger>=2.0.0",

    # Optional integrations
    "gspread>=6.0.0",
    "google-auth>=2.27.0",
    "python-telegram-bot>=20.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.20.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["api", "crawlers", "services", "storage", "parsers"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
